requirements.txt
typer
python-dotenv
pydantic

langchain
langchain-openai

chromadb
sentence-transformers

tqdm
rich
hf_xet